♻️refact: Formatt - Maior manutenabilidade do código - Dependency Injection - Maior Granunalidade
