+++ TheatricalPlayersRefactoringKata/Domain/Interfaces/IStatementFormatter.cs
@@ -1,0 +1,8 @@
+using TheatricalPlayersRefactoringKata.Domain.Entities;
+
+namespace TheatricalPlayersRefactoringKata.Domain.Interfaces;
+
+public interface IStatementFormatter
+{
+    string Format(Invoice invoice);
+}
